requirements.txt
python-telegram-bot==20.0
google-cloud-translate
flask
requests
